Constrain child windows to monitor bounds
If a parent window is at a corner of a desktop, and it is smaller than
a child window, it is possible to end up in a situation where the child
window is positioned outside of visible and reachable area.
Coupled with the fact that locations and dimensions of most windows
get persisted this may lead to a situation where the window may become
permanently inaccessible until the settings are reset.